chore: rename marquee-components and folders
