+++ src/atoms/marquee-content/MarqueeContent.jsx
